サイト別adapter構成へリファクタ
- 変更理由: 共通ロジックとサイト固有ロジックを分離し、対応サイトを追加しやすくするため
- 変更内容: base/chatgpt/gemini adapter と DOM ヘルパーを追加し、composer/send/content を委譲構成へ整理
- 動作確認: npm run build
- 動作確認: npm run test:e2e（権限昇格で実行、2件通過）
