+++ src/types.ts
@@ -3,4 +3,12 @@ namespace AwesomeInput {
     | HTMLTextAreaElement
     | HTMLInputElement
     | HTMLElement;
+
+  export interface SiteAdapter {
+    id: string;
+    matches(hostname: string): boolean;
+    findComposer(): EditableElement | null;
+    findSendButton(composer: EditableElement | null): HTMLButtonElement | null;
+    insertNewline(composer: EditableElement): boolean;
+  }
 }
