+++ src/utils.ts
@@ -1,14 +1,6 @@
 namespace AwesomeInput {
-  export function sleep(ms: number): Promise<void> {
-    return new Promise((resolve) => window.setTimeout(resolve, ms));
-  }
-
   export function currentHostname(): string {
     return window.location.hostname.toLowerCase();
-  }
-
-  export function isGeminiHost(): boolean {
-    return currentHostname() === "gemini.google.com";
   }
 
   export function isMac(): boolean {
@@ -23,10 +15,5 @@ namespace AwesomeInput {
     }
 
     return event.ctrlKey && !event.metaKey;
-  }
-
-  export function prefersSyntheticLineBreak(): boolean {
-    const hostname = currentHostname();
-    return hostname === "chatgpt.com" || hostname === "chat.openai.com";
   }
 }
